Create README.md
Common Approach has developed code for a bare-bones impact measurement software. Originally developed for our own testing purposes, we are sharing it here in hopes it will be useful to anyone developing or expanding the functionality of software related to impact measurement and management. It provides a starter kit for building impact measurement software aligned with the Common Impact Data Standard (see link below for more info). 

This repository contains demo code to import and view data impact data aligned with the Common Impact Data Standard. Documentation for the code can be accessed through the Readme file or on the Wiki page. 
i